# build: faster newLinesIndices via bytes.IndexByte and buffer re-use