See the individual commits for an explanation. The overall performance improvement of BenchmarkTagsToSections on my machine is:

```
goos: linux
goarch: amd64
pkg: github.com/sourcegraph/zoekt/build
cpu: AMD Ryzen 9 5950X 16-Core Processor

old time/op    new time/op    delta
   188µs ± 7%     101µs ± 3%  -46.10%  (p=0.000 n=10+10)

old alloc/op   new alloc/op   delta
  79.3kB ± 0%    36.3kB ± 0%  -54.24%  (p=0.000 n=9+10)

old allocs/op  new allocs/op  delta
     443 ± 0%       441 ± 0%   -0.45%  (p=0.000 n=10+10)
```