@keegancsmith 2023-11-06
This is quite fast, so I don't this the speed improvements are that important. However, the memory usage improvements are likely noticeable due to less work the GC needs to do.

Also I recently instrumented the time taken by the ctags parser and didn't include the time taken by tagsToSections. I think given the speed of this that is likely the correct call. If we are slow here it will be ctags not this conversion code.